refactor(Projectile): Merge into GameObject.hpp
