+++ pvzclass/Classes.hpp
@@ -4,7 +4,6 @@
 #include "Classes/ZenGarden.hpp"
 
 #include "Classes/GameObject.hpp"
-#include "Classes/Projectile.hpp"
 #include "Classes/LawnMower.hpp"
 #include "Classes/Griditem.hpp"
 #include "Classes/Mouse.hpp"
