+++ pvzclass/Classes/Griditem.cpp
@@ -1,6 +1,5 @@
 #include "../PVZ.h"
 #include "GameObject.hpp"
-#include "Projectile.hpp"
 #include "Griditem.hpp"
 
 PVZ::Griditem::Griditem(int indexoraddress)
