+++ pvzclass/pvzclass.vcxproj.filters
@@ -854,9 +854,6 @@
     <ClInclude Include="Classes\LawnMower.hpp">
       <Filter>头文件\Classes</Filter>
     </ClInclude>
-    <ClInclude Include="Classes\Projectile.hpp">
-      <Filter>头文件\Classes</Filter>
-    </ClInclude>
     <ClInclude Include="Classes\Griditem.hpp">
       <Filter>头文件\Classes</Filter>
     </ClInclude>
